#22: checking out the branch, not the PR remote (head would be in a detached state)
This is because for scalafmt we need to find the 'base' branch - which is master in this case, and compare the list of changed files against it, in order to check the formatting against such newly changed files only. Also, 'origin/' isn't needed because the 'base' branch will be 'origin/master', not just 'master' - see file build.sbt, line starting with `scalafmtFilter.withRank` and the expression starting with `git show-branch -a`.